requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.7
gunicorn==21.2.0
idna==3.4
instagrapi==2.0.0
kombu==5.3.3
packaging==23.2
Pillow==10.1.0
prompt-toolkit==3.0.39
psycopg2==2.9.9
pycryptodomex==3.18.0
pydantic==1.10.9
PySocks==1.7.1
python-dateutil==2.8.2
python-magic==0.4.27
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.9
moviepy>=1.0.3
django-dirtyfields==1.9.2